test coverage, bugfix for summarizer
